requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-material-admin==1.8.6
djangorestframework==3.14.0
gunicorn==21.2.0
gevent ==24.2.1
idna==3.6
mysqlclient==2.2.4
packaging==24.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
